+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Option to disable processing of "style" tags. [#45](https://github.com/Stranger6667/css-inline/issues/45)
+- Option to inline additional CSS. [#45](https://github.com/Stranger6667/css-inline/issues/45)
 
 ## [0.3.3] - 2020-07-07